+++ CHANGELOG.md
@@ -6,3 +6,4 @@ All notable changes to the "zotero-for-typst" extension will be documented in th
 
 - Insert citation as `#cite(...)` command.
 - Setting for host and port.
+- Insert citation when typing `ZCT` (optional).